services/monitoring: Fix firewall interface
The monitoring network interface has a different name on hamilton
compared to the others. This caused firewall to allowlist prometheus and
alertmanager on the wrong interface, thus hamilton appearing down to the
other hosts.

By adding a `cj.monitoring` module (just basic variable storage) we can
define the interface in the host specific configuration and use it for
e.g. the firewall