services/monitoring: Change monitoring concept
Previously, each host was monitorig itsetlf. The other hosts were only
checking if there were no problems with the alertmanager connection and
so on.
Tho in case a host goes semi-down (e.g. disk running full or very low
memory situation) it may not be able to send alerts about it's current
situation. And in those cases, it's possible that the other signs (e.g.
failed notifications from alertmanager) may not or only with a delay
indicate this issue.

With this change, all hosts monitor the node exporters of all hosts.
That way, even if a host has no space left on disk, other hosts will
detect this situation. To avoid duplicate alerts (which were already
happening with e.g. unavailable alertmanagers which would be
detected and alerted separately by both hosts that were fine), the
alertmanagers now form a cluster. Alertmanager automatically takes care
of deduplicating the incomming notifications.