Fix pysaml2 build
Apply the patch from the PR that replaces pyopenssl with cryptography.
The PR has not been reviewed by upstream, so there is a chance of
something being not right. However, the tests pass. So it should
probably be allright?

(And in turn we can finally update matrix and the rest of our services
and close some very real CVEs)